refactor: remove deprecated story files and create new catalog structure for atoms
- Deleted NoteEditorMobile, CollectionsPage, JournalPageShell, Syntax, Atoms stories.
- Created new catalog structure for atom components including Button, CalendarDatePicker, Card, Dialog, DropdownMenu, and WodPlaygroundButton.
- Each atom story showcases variants, sizes, states, and interactive examples where applicable.
